refactor(contract): move foreign-chain support and provider policy into api/foreign_chain_support.rs
